refactored component.rs and added Program struct
