+++ src/welcome_message.rs
@@ -1,0 +1,31 @@
+use crate::configuration_file;
+
+fn print_rust_ascii() {
+    let ascii = "
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░▒░░░░░░░░░░░░░░░░░░░░░░░▓
+░░░░░░░░░░░░░░░░░░░░░░░░░░▒░░▓▓▓▒▓▓▓▒▒▓▓░░▒▒░░░░░░░░░░░░░░▓
+░░░░░░░░░░░░░░░░░░░░░░▒░░▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓░░░░░░░░░░░░░░▓
+░░░░░░░░░░░░░░░░░░░░░░▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▒░░░░░░░░░░▓
+░░░░░░░░░░░░░░░░░░▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓░░░░░░░▓
+░░░░░░░░░░░░░░░░░░▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓░░░░░░░▓
+░░░░░░░░░░░░░░░▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▒░░░░▓
+░░░░░▒▓▓░░░░░░░░▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓░░░░░▓
+▒▓▓▓▒▒▓▓▓▒░░░▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓▒▒██▓▓▓▓▓▓▓▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓░░▓
+░▓▓▓▓▓▓▓▓▒░░░░▒▓▓▓▓▓▓▓▓▓▓▓▓▓░░░██▓▓▓▓█░░▓██▓▓▓▓▓▓▓▓▓▓▓▓░░░▓
+░░▒▓▓▓▓▓▓▓▒░▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓██████▓▓▓▓██████▓▓▓▓▓▓▓▓▓▓▓▓▓▒░▓
+░░░░░▒▒░░▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓████▓▓▓▓▓████▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▒▓
+░░░░░░░░░░░░▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓░▓
+░░░░░░░░░░░░░▒▓▓▓▓▒▓▓░▒▒▒▒▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▒▓▓▓▒▒▓▓▓▓▓▓▒░▓
+░░░░░░░░░░░░░░▒▓▓▓▒░▓░░░░░░░░░░░░░░░░▒▓▓▓▓▓▓▓▓▓▒░▓▒░▒▓▓░░░▓
+░░░░░░░░░░░░░░░░▒▓▒░░░░░░░░░░░░░░░░░▒▓▓▓▓▓▓▓▓▒░░░░░░▓▓░░░░▓
+░░░░░░░░░░░░░░░░░░▒░░░░░░░░░░░░░░░░░▒▒░░░▓▓▓▓░░░░░░░▒░░░░░▓
+░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░░▓▓▓▒░░░░░░░░░░░░░░";
+    println!("{}", ascii)
+}
+
+
+pub fn display_welcome_message() {
+    print_rust_ascii();
+    let message = configuration_file::get_welcome_message();
+    println!("{}", message)
+}
